remove randomize desktop and mobile page component
